refactor detection of extended kb to own service
